pyproject.toml
[project]
name = "cdn-digital-ocean-image"
version = "0.1.0"
description = ""
authors = [
    {name = "OliverRoat",email = "oliver.roat@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "boto3 (>=1.37.18,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
